# docs/openapi-telemetry.yaml
openapi: 3.0.3
info:
  title: Telemetry Service API
  version: 1.0.0
servers:
  - url: http://localhost:8085/telemetry
paths:
  /records:
    get:
      summary: List consumption records
      responses:
        '200':
          description: Consumption records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsumptionRecord'
    post:
      summary: Create a consumption record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRecordRequest'
      responses:
        '201':
          description: Created record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumptionRecord'
  /records/{id}:
    get:
      summary: Get record by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumptionRecord'
  /records/vehicle/{vehicleId}:
    get:
      summary: List records for a vehicle
      parameters:
        - in: path
          name: vehicleId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vehicle telemetry records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsumptionRecord'
  /anomalies:
    get:
      summary: List consumption anomalies
      responses:
        '200':
          description: Anomalies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsumptionAnomaly'
components:
  schemas:
    ConsumptionRecord:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: string
        metricName:
          type: string
        expectedDays:
          type: integer
        actualDays:
          type: integer
        lastCheckedDate:
          type: string
          format: date
        anomaly:
          type: boolean
    ConsumptionRecordRequest:
      type: object
      required: [vehicleId, metricName, expectedDays, actualDays, lastCheckedDate]
      properties:
        vehicleId:
          type: string
        metricName:
          type: string
        expectedDays:
          type: integer
        actualDays:
          type: integer
        lastCheckedDate:
          type: string
          format: date
    ConsumptionAnomaly:
      type: object
      properties:
        recordId:
          type: integer
        vehicleId:
          type: string
        metricName:
          type: string
        expectedDays:
          type: integer
        actualDays:
          type: integer
        deviationDays:
          type: integer